// types/enums.ts
export enum PaymentMethodIconEnum {
	Card = 'credit-card-outline',
	Cash = 'cash',
	Transfer = 'bank-transfer',
}

export enum PaymentMethodEnum {
	Card = 'card',
	Cash = 'cash',
	Transfer = 'transfer',
}

export enum TransactionTypeEnum {
	Income = 'income',
	Expense = 'expense',
}

export enum TransactionCategoryEnum {
	Food = 'food',
	Transport = 'transport',
	Health = 'health',
	Education = 'education',
	Entertainment = 'entertainment',
    Empty = '',
	Other = 'other',
}